+++ HISTORY.md
@@ -15,6 +15,7 @@
 * blocks: refactored adding and removing inputs in variadic slots
 * blocks: adjusted inserting / deleting single inputs in variadic infix slots
 * objects: adjusted block-search-bar for variadic infix reporters
+* objects: adapted formula editor for variadic infix reporters
 
 ### 2022-02-28
 * blocks, objects, threads, store: made addition reporter variadic